backend/requirements.txt
Django==5.2.10
djangorestframework==3.16.1
psycopg==3.3.2